Refactor notifications module for improved file organization and clarity
- Restructured the notifications module by creating dedicated `/services` and `/processors` directories to enhance code organization.
- Updated import paths in `notifications.module.ts` and `notifications.service.ts` to reflect the new structure.
- Introduced index files for both services and processors to streamline imports.
- Maintained existing functionality while improving the overall clarity and maintainability of the codebase.
